requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.0.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.2
pytest-flask==1.2.0
python-dateutil==2.8.2
email-validator==2.0.0